+++ src/main/java/app/domains/asset/model/AssetFilterOptionsDto.java
@@ -1,4 +1,4 @@
-package app.domains.assets.model;
+package app.domains.asset.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
