+++ src/main/java/app/domains/asset/model/AssetPartsUpdateDto.java
@@ -1,4 +1,4 @@
-package app.domains.assets.model;
+package app.domains.asset.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,7 +11,10 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class AssetPartsDto {
+public class AssetPartsUpdateDto {
     private Asset asset;
     private List<Part> parts;
+    private boolean assetCycleChanged;
+    private boolean partCyclesChanged;
+    private boolean partCountChanged;
 }
